Menajo de errores en la vista de las tareas

+++ app/Http/Controllers/WEB/TodoController.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\WEB;
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use App\Models\Todo;
+use Illuminate\Http\Request;
 
 class TodoController extends Controller
 {
@@ -21,7 +22,7 @@ class TodoController extends Controller
      */
     public function create()
     {
-        //
+        return to_route('todos');
     }
 
     /**
@@ -29,14 +30,14 @@ class TodoController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        $validateData = $request->validate([
             'title' => 'required|min:3',
             'date' => 'nullable|date'
         ]);
 
         $todo = new Todo;
-        $todo->title = $request->title;
-        $todo->date = $request->date;
+        $todo->title = $validateData['title'];
+        $todo->date = $request['date'];
 
         $todo->save();
 
@@ -48,6 +49,7 @@ class TodoController extends Controller
      */
     public function show(string $id)
     {
+        return to_route('todos');
     }
 
     /**
@@ -59,6 +61,10 @@ class TodoController extends Controller
         $todos = Todo::all();
         $todo = Todo::find($id);
 
+        if (is_null($todo)) {
+            return to_route('todos')->with('error', 'Tarea no encontrada');
+        }
+
         return view('todos.edit', ['todo' => $todo, 'todos' => $todos]);
     }
 
@@ -68,8 +74,18 @@ class TodoController extends Controller
     public function update(Request $request, string $id)
     {
         $todo = Todo::find($id);
-        $todo->title = $request->title;
-        $todo->date = $request->date;
+
+        if (is_null($todo)) {
+            return to_route('todos')->with('error', 'Tarea no encontrada');
+        }
+
+        $validateData = $request->validate([
+            'title' => 'required|min:3',
+            'date' => 'nullable|date'
+        ]);
+
+        $todo->title = $validateData['title'];
+        $todo->date = $request['date'];
         $todo->save();
 
         return to_route('todos')->with('success', 'Tarea Actualizada');
@@ -81,6 +97,9 @@ class TodoController extends Controller
     public function destroy(string $id)
     {
         $todo = Todo::find($id);
+        if (is_null($todo)) {
+            return to_route('todos')->with('error', 'Tarea no encontrada');
+        }
         $todo->delete();
 
         return to_route('todos')->with('success', 'Tarea Eliminada');
